Few more fixes for http://qa-jenkins.gainsight.com:8080/job/Cockpit_A…

+++ src/com/gainsight/sfdc/sfWidgets/oppWidget/tests/OppWidget_CockpitTests.java
@@ -338,8 +338,9 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		        	if(task.getAssignee()==null) {
 		                task.setAssignee(sfdcInfo.getUserFullName());
 		            }
-		        	task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		        	task.setFromCustomer360orWidgets(true);
+
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate()))); 
+					task.setFromCustomer360orWidgets(true);
 		        	}
 		       oppWfPage  = oppWfPage.applyPlayBook(cta, testData.get("Playbook"), tasks,true);
 		       for(Task task : tasks) {
@@ -365,8 +366,9 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		        	if(task.getAssignee()==null) {
 		                task.setAssignee(sfdcInfo.getUserFullName());
 		            }
-		        	task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		        	task.setAssignee(users[i]); if(++i >=5) i=0;
+
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate()))); 
+					task.setAssignee(users[i]); if(++i >=5) i=0;
 		        	task.setFromCustomer360orWidgets(true);
 		        	}
 
@@ -392,8 +394,9 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		        	if(task.getAssignee()==null) {
 		                task.setAssignee(sfdcInfo.getUserFullName());
 		            }
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate()))); 
+					task.setFromCustomer360orWidgets(true);
 		        	}
 		        
 		        //Applying Playbook and verifying tasks
@@ -408,8 +411,8 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		          	if(task.getAssignee()==null) {
 		                task.setAssignee(sfdcInfo.getUserFullName());
 		            }
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		          	}
 		       cta.setDueDate(getHighestTaskDate(tasks));
 		       oppWfPage = oppWfPage.applyPlayBook(cta, testData.get("UpdatedPlaybook"), updatedTasks,false);
@@ -437,8 +440,9 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		       ArrayList<Task> tasks  = getTaskFromSFDC(testData.get("Playbook"));
 		        for(Task task : tasks) {
 		        	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate()))); 
+					task.setFromCustomer360orWidgets(true);
 		        	}
 
 		       oppWfPage  = oppWfPage.applyPlayBook(cta, testData.get("Playbook"), tasks,true);
@@ -461,8 +465,9 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		       ArrayList<Task> tasks  = getTaskFromSFDC(testData.get("Playbook"));
 		        for(Task task : tasks) {
 		        	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate()))); 
+					task.setFromCustomer360orWidgets(true);
 		        	}
 		        
 		        //Applying Playbook and verifying tasks
@@ -475,8 +480,8 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		       ArrayList<Task> updatedTasks = getTaskFromSFDC(testData.get("UpdatedPlaybook"));
 		       for(Task task : updatedTasks) {
 		          	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		           task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		           task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		          	}
 		       cta.setDueDate(getHighestTaskDate(tasks));
 		       oppWfPage = oppWfPage.applyPlayBook(cta, testData.get("UpdatedPlaybook"), updatedTasks,false);
@@ -504,8 +509,8 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		       ArrayList<Task> tasks  = getTaskFromSFDC(testData.get("Playbook"));
 		       for(Task task : tasks) {
 		        	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		        	}
 		        
 		       oppWfPage = oppWfPage.applyPlayBook(cta, testData.get("Playbook"), tasks,true);
@@ -528,8 +533,8 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		       ArrayList<Task> tasks  = getTaskFromSFDC(testData.get("Playbook"));
 		       for(Task task : tasks) {
 		        	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		           task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		           task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		        	}
 		        //Applying Playbook and verifying tasks
 		       oppWfPage  = oppWfPage.applyPlayBook(cta, testData.get("Playbook"), tasks,true);
@@ -541,8 +546,8 @@ public class OppWidget_CockpitTests  extends WorkflowSetup {
 		       ArrayList<Task> updatedTasks = getTaskFromSFDC(testData.get("UpdatedPlaybook"));
 		       for(Task task : updatedTasks) {
 		          	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		           task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		           task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		          	}
 		       cta.setDueDate(getHighestTaskDate(tasks));
 		       oppWfPage = oppWfPage.applyPlayBook(cta, testData.get("UpdatedPlaybook"), updatedTasks,false);
